# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for saving movies",
        "title": "CRUD movie manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/movies": {
            "get": {
                "description": "Get all movies list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get movies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Movie"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add new movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add new movie",
                "parameters": [
                    {
                        "description": "Add movie to list, 'id' and 'savedAt' not necessary params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MovieMainInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MovieMainInfo"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.MovieMainInfo"
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "Get movies by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get movies by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Movie"
                        }
                    }
                }
            },
            "put": {
                "description": "Update movie info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update movie info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Add movie to list, 'id' and 'savedAt' not necessary params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MovieMainInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MovieMainInfo"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteMovie from list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "DeleteMovie from list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transport.statusResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Movie": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "release": {
                    "type": "string"
                },
                "savedAt": {
                    "type": "string"
                },
                "streamingService": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "domain.MovieMainInfo": {
            "type": "object",
            "properties": {
                "release": {
                    "type": "string"
                },
                "streamingService": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "transport.statusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}